fix: fix type issues
